Implementación de imagenes finalizado, pasará a pruebas

+++ src/infraestructure/repositories/evaluation-details-repository.ts
@@ -1,13 +1,18 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { EvaluationDetailEntity } from 'src/application-core/domain/entities/evaluation-detail-entity';
+import { EvaluationEntity } from 'src/application-core/domain/entities/evaluation-entity';
 import { EvaluationDetailRequestDto } from 'src/application-core/dto/requests/evaluation-detail-dto';
+import { NotFoundException } from 'src/application-core/exception/not-found-exception';
+import { ValidationException } from 'src/application-core/exception/validation-exception';
 import IEvaluationDetail from 'src/application-core/interfaces/i-evaluation-detail';
 import { Repository } from 'typeorm';
 
 @Injectable()
 export class EvaluationDetailsRepository implements IEvaluationDetail {
   constructor(
+    @InjectRepository(EvaluationEntity)
+    private readonly evaluationRepository: Repository<EvaluationEntity>,
     @InjectRepository(EvaluationDetailEntity)
     private readonly evaluationDetailRepository: Repository<EvaluationDetailEntity>,
   ) {}
@@ -38,7 +43,24 @@ export class EvaluationDetailsRepository implements IEvaluationDetail {
 
   async saveEvaluationDetail(
     evaluationDetail: EvaluationDetailRequestDto,
+    username: string,
   ): Promise<EvaluationDetailEntity> {
+    const evaluationExists = await this.evaluationRepository.findOne({
+      where: { evaluationId: evaluationDetail.evaluationId },
+    });
+    if (!evaluationExists) {
+      throw new NotFoundException(
+        'La evaluación no existe, por favor verifique el ID de la evaluación',
+      );
+    }
+    if (!evaluationExists.isOpen) {
+      throw new ValidationException('La evaluación no se encuentra abierta');
+    }
+    if (evaluationExists.username !== username) {
+      throw new ValidationException(
+        'No tiene permiso para agregar detalles a esta evaluación',
+      );
+    }
     return await this.evaluationDetailRepository.save({
       evaluation: { evaluationId: evaluationDetail.evaluationId },
       category: { nKey: evaluationDetail.categoryId },
@@ -53,6 +75,7 @@ export class EvaluationDetailsRepository implements IEvaluationDetail {
   async updateEvaluationDetail(
     id: number,
     evaluationDetail: EvaluationDetailRequestDto,
+    username: string,
   ): Promise<EvaluationDetailEntity> {
     const existingEvaluationDetail =
       await this.evaluationDetailRepository.findOne({
@@ -68,16 +91,35 @@ export class EvaluationDetailsRepository implements IEvaluationDetail {
       ...existingEvaluationDetail,
       ...evaluationDetail,
     };
+    if (!existingEvaluationDetail) {
+      throw new NotFoundException('Detalle de evaluación no encontrado');
+    }
+    if (existingEvaluationDetail.evaluation.username !== username) {
+      throw new ValidationException(
+        'No tiene permiso para actualizar este detalle de evaluación',
+      );
+    }
+    if (!existingEvaluationDetail.evaluation.isOpen) {
+      throw new ValidationException('La evaluación no se encuentra abierta');
+    }
     return await this.evaluationDetailRepository.save(updatedEvaluationDetail);
   }
 
-  async deleteEvaluationDetail(id: number): Promise<string> {
+  async deleteEvaluationDetail(id: number, username: string): Promise<string> {
     const existingEvaluationDetail =
       await this.evaluationDetailRepository.findOne({
         where: { evaluationDetailId: id },
       });
     if (!existingEvaluationDetail) {
-      return 'No se encontró el detalle de evaluación';
+      throw new NotFoundException('No se encontró el detalle de evaluación');
+    }
+    if (existingEvaluationDetail.evaluation.username !== username) {
+      throw new ValidationException(
+        'No tiene permiso para eliminar este detalle de evaluación',
+      );
+    }
+    if (!existingEvaluationDetail.evaluation.isOpen) {
+      throw new ValidationException('La evaluación no se encuentra abierta');
     }
     await this.evaluationDetailRepository.delete(id);
     return 'Detalle de evaluación eliminado correctamente';
